# frontend/docs/openapi-documentation.json
{"openapi":"3.1.0","info":{"title":"SmartAnalyticsApp API","version":"0.1.0"},"paths":{"/files/upload":{"post":{"tags":["files"],"summary":"Upload Data File","description":"Handle file uploads, saving them to disk and recording metadata in memory.\nArgs:\n    file: The uploaded file, provided as a form-data field.\nReturns:\n    An UploadResponseDTO containing metadata about the uploaded file.\nRaises:\n    HTTPException: If the file is missing, cannot be saved, or if the filename is","operationId":"upload_data_file_files_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_data_file_files_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/files/metadata/{upload_id}":{"get":{"tags":["files"],"summary":"Get File Metadata","description":"Retrieve metadata for a specific uploaded file by its upload ID.\nArgs:\n    upload_id: The unique identifier for the uploaded file.\nReturns:\n    An UploadResponseDTO containing metadata about the specified file.\nRaises:\n    HTTPException: If no file metadata is found for the given upload ID.","operationId":"get_file_metadata_files_metadata__upload_id__get","parameters":[{"name":"upload_id","in":"path","required":true,"schema":{"type":"string","title":"Upload Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/files/metadata":{"get":{"tags":["files"],"summary":"List File Metadata","description":"Retrieve metadata for all uploaded files in the current session.\nReturns:\n    A list of UploadResponseDTO objects, each containing metadata about an uploaded file.","operationId":"list_file_metadata_files_metadata_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UploadResponseDTO"},"type":"array","title":"Response List File Metadata Files Metadata Get"}}}}}}},"/files/diagrams/{upload_id}/{diagram_type}":{"get":{"tags":["files"],"summary":"List Diagrams","description":"Retrieve diagram images for a specific upload ID and diagram type.","operationId":"list_diagrams_files_diagrams__upload_id___diagram_type__get","parameters":[{"name":"upload_id","in":"path","required":true,"schema":{"type":"string","title":"Upload Id"}},{"name":"diagram_type","in":"path","required":true,"schema":{"type":"string","title":"Diagram Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagramsResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preprocessing/type/{upload_id}":{"get":{"tags":["preprocessing"],"summary":"Get Preprocessing Method Types","description":"Get a list of available preprocessing method types.\nArgs:\n    upload_id: The unique identifier of the uploaded file for which to determine preprocessing methods.\nreturns:\n    A list of MethodTypeResponseDTO objects representing the supported preprocessing methods.","operationId":"get_preprocessing_method_types_preprocessing_type__upload_id__get","parameters":[{"name":"upload_id","in":"path","required":true,"schema":{"type":"string","title":"Upload Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MethodTypeResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preprocessing/run":{"post":{"tags":["preprocessing"],"summary":"Run Preprocessing","description":"Run the preprocessing code for a given uploaded file.\nArgs:\n    request: The request object containing the upload ID and method type.\nreturns:\n    A dictionary containing the status of the preprocessing operation.","operationId":"run_preprocessing_preprocessing_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreprocessingRunRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreprocessingRunResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clustering/run":{"post":{"tags":["clustering"],"summary":"Run Clustering","description":"Run a clustering algorithm on the uploaded data.\nArgs:\n    request: ClusteringRunRequestDTO containing the upload ID and algorithm.\nReturns:\n    ClusteringRunResponseDTO with the output file path, algorithm used, number of clusters, and noise points (if applicable).","operationId":"run_clustering_clustering_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClusteringRunRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClusteringRunResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clustering/best":{"post":{"tags":["clustering"],"summary":"Run Best Clustering","description":"Run all the Clustering algorithms\n\nArgs:\n    request (BestClusteringRunRequestDTO): The id of the upload to run the algorithms on.\n\nReturns:\n    BestClusteringRunResponseDTO: The best clustering results.","operationId":"run_best_clustering_clustering_best_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BestClusteringRunRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BestClusteringRunResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/association/run":{"post":{"tags":["Association Rule Mining"],"summary":"Run Association Rule Mining","description":"Run association rule mining on the uploaded data.\n\nArgs:\n    request (AssociationRunRequestDTO): The request data containing upload ID and thresholds.\n\nReturns:\n    AssociationRunResponseDTO: The response containing paths to the results files.","operationId":"run_association_rule_mining_association_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssociationRunRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssociationRunResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pca/run":{"post":{"tags":["PCA"],"summary":"Run Pca","description":"Run PCA on the uploaded data.\n\nArgs:\n    request (PCARunRequestDTO): The request data containing upload ID and number of components.\n\nReturns:\n    PCARunResponseDTO: The response containing the path to the PCA result file.","operationId":"run_pca_pca_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PCARunRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PCARunResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/time-series/run":{"post":{"tags":["time_series"],"summary":"Run Time Series","description":"Run time series forecasting for an uploaded file.","operationId":"run_time_series_time_series_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSeriesRunRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSeriesRunResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/generate":{"post":{"tags":["reports"],"summary":"Generate Report","operationId":"generate_report_reports_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportResponseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","description":"Root endpoint to verify the API is running.\nreturns:\n    A simple status message.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Read Root  Get"}}}}}}},"/session":{"get":{"summary":"Get Session Summary","description":"Get a summary of the current session, including all uploaded files.\nreturns:\n    A dictionary mapping upload IDs to their metadata.","operationId":"get_session_summary_session_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"$ref":"#/components/schemas/UploadResponseDTO"},"type":"object","title":"Response Get Session Summary Session Get"}}}}}}}},"components":{"schemas":{"AlgorithmReportInputDTO":{"properties":{"name":{"type":"string","title":"Name"},"measures":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Measures"},"params":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Params"},"outputs":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Outputs"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["name"],"title":"AlgorithmReportInputDTO"},"AssociationRunRequestDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"},"min_support":{"type":"number","title":"Min Support"},"min_confidence":{"type":"number","title":"Min Confidence"},"min_lift":{"type":"number","title":"Min Lift"}},"type":"object","required":["upload_id","min_support","min_confidence","min_lift"],"title":"AssociationRunRequestDTO"},"AssociationRunResponseDTO":{"properties":{"frequent_itemsets_file_path":{"type":"string","title":"Frequent Itemsets File Path"},"association_rules_file_path":{"type":"string","title":"Association Rules File Path"}},"type":"object","required":["frequent_itemsets_file_path","association_rules_file_path"],"title":"AssociationRunResponseDTO"},"BestClusteringRunRequestDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"}},"type":"object","required":["upload_id"],"title":"BestClusteringRunRequestDTO"},"BestClusteringRunResponseDTO":{"properties":{"output_file_path":{"type":"string","title":"Output File Path"},"best_algorithm":{"$ref":"#/components/schemas/ClusteringAlgorithm"},"results":{"items":{"$ref":"#/components/schemas/ClusteringScoreDTO"},"type":"array","title":"Results"}},"type":"object","required":["output_file_path","best_algorithm","results"],"title":"BestClusteringRunResponseDTO"},"Body_upload_data_file_files_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_data_file_files_upload_post"},"ClusteringAlgorithm":{"type":"string","enum":["kmeans","dbscan","hierarchical"],"title":"ClusteringAlgorithm"},"ClusteringRunRequestDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"},"algorithm":{"$ref":"#/components/schemas/ClusteringAlgorithm"}},"type":"object","required":["upload_id","algorithm"],"title":"ClusteringRunRequestDTO"},"ClusteringRunResponseDTO":{"properties":{"output_file_path":{"type":"string","title":"Output File Path"},"algorithm":{"$ref":"#/components/schemas/ClusteringAlgorithm"},"n_clusters":{"type":"integer","title":"N Clusters"},"noise_points":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Noise Points"}},"type":"object","required":["output_file_path","algorithm","n_clusters"],"title":"ClusteringRunResponseDTO"},"ClusteringScoreDTO":{"properties":{"algorithm":{"$ref":"#/components/schemas/ClusteringAlgorithm"},"silhouette":{"type":"number","title":"Silhouette"},"n_clusters":{"type":"integer","title":"N Clusters"},"noise_points":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Noise Points"}},"type":"object","required":["algorithm","silhouette","n_clusters"],"title":"ClusteringScoreDTO"},"DiagramImageDTO":{"properties":{"filename":{"type":"string","title":"Filename"},"content_type":{"type":"string","title":"Content Type"},"data_base64":{"type":"string","title":"Data Base64"}},"type":"object","required":["filename","content_type","data_base64"],"title":"DiagramImageDTO"},"DiagramsResponseDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"},"diagram_type":{"type":"string","title":"Diagram Type"},"images":{"items":{"$ref":"#/components/schemas/DiagramImageDTO"},"type":"array","title":"Images"}},"type":"object","required":["upload_id","diagram_type","images"],"title":"DiagramsResponseDTO"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MethodType":{"type":"string","enum":["clustering","association_rule","pca","time_series"],"title":"MethodType"},"MethodTypeResponseDTO":{"properties":{"method_type":{"items":{"$ref":"#/components/schemas/MethodType"},"type":"array","title":"Method Type"}},"type":"object","required":["method_type"],"title":"MethodTypeResponseDTO"},"PCAMethodType":{"type":"string","enum":["SVD","covariance_matrix"],"title":"PCAMethodType"},"PCARunRequestDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"},"n_components":{"type":"number","title":"N Components","default":0.95},"threshold":{"type":"number","title":"Threshold","default":0.95}},"type":"object","required":["upload_id"],"title":"PCARunRequestDTO"},"PCARunResponseDTO":{"properties":{"transformed_data_file_path":{"type":"string","title":"Transformed Data File Path"},"explained_variance_ratio":{"items":{},"type":"array","title":"Explained Variance Ratio"},"cumulative_variance":{"items":{},"type":"array","title":"Cumulative Variance"},"pca_method_type":{"$ref":"#/components/schemas/PCAMethodType"}},"type":"object","required":["transformed_data_file_path","explained_variance_ratio","cumulative_variance","pca_method_type"],"title":"PCARunResponseDTO"},"PreprocessingRunRequestDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"},"method_type":{"$ref":"#/components/schemas/MethodType"}},"type":"object","required":["upload_id","method_type"],"title":"PreprocessingRunRequestDTO"},"PreprocessingRunResponseDTO":{"properties":{"output_file_path":{"type":"string","title":"Output File Path"}},"type":"object","required":["output_file_path"],"title":"PreprocessingRunResponseDTO"},"ReportRequestDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"},"algorithms":{"items":{"$ref":"#/components/schemas/AlgorithmReportInputDTO"},"type":"array","title":"Algorithms"}},"type":"object","required":["upload_id","algorithms"],"title":"ReportRequestDTO"},"ReportResponseDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"},"report_text":{"type":"string","title":"Report Text"},"schema":{"additionalProperties":true,"type":"object","title":"Schema"}},"type":"object","required":["upload_id","report_text","schema"],"title":"ReportResponseDTO"},"TimeSeriesMethodType":{"type":"string","enum":["linear_regression","arima"],"title":"TimeSeriesMethodType"},"TimeSeriesRunRequestDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"},"method_type":{"$ref":"#/components/schemas/TimeSeriesMethodType","default":"linear_regression"}},"type":"object","required":["upload_id"],"title":"TimeSeriesRunRequestDTO"},"TimeSeriesRunResponseDTO":{"properties":{"transformed_data_file_path":{"type":"string","title":"Transformed Data File Path"},"forecast_file_path":{"type":"string","title":"Forecast File Path"},"historical_plot_path":{"type":"string","title":"Historical Plot Path"},"forecast_plot_path":{"type":"string","title":"Forecast Plot Path"},"target_column":{"type":"string","title":"Target Column"},"datetime_column":{"type":"string","title":"Datetime Column"},"metrics":{"additionalProperties":true,"type":"object","title":"Metrics"},"time_series_method_type":{"$ref":"#/components/schemas/TimeSeriesMethodType"}},"type":"object","required":["transformed_data_file_path","forecast_file_path","historical_plot_path","forecast_plot_path","target_column","datetime_column","metrics","time_series_method_type"],"title":"TimeSeriesRunResponseDTO"},"UploadResponseDTO":{"properties":{"upload_id":{"type":"string","title":"Upload Id"},"filename":{"type":"string","title":"Filename"},"content_type":{"type":"string","title":"Content Type"},"bytes":{"type":"integer","title":"Bytes"},"path":{"type":"string","title":"Path"},"method_types":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Method Types"}},"type":"object","required":["upload_id","filename","content_type","bytes","path"],"title":"UploadResponseDTO","description":"Data Transfer Object for responses related to file uploads.\nAttributes:\n    upload_id: A unique identifier for the uploaded file.\n    filename: The original filename of the uploaded file.\n    content_type: The MIME type of the uploaded file.\n    bytes: The size of the uploaded file in bytes."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}